+++ lib/drivers/mongo/build-mongo-execute.js
@@ -1,5 +1,6 @@
 const camelobj = require('camelobj');
-const connectMongo = require('./../../../test/data/fixtures/connect-mongo');
+
+const connectMongo = require('./util/connect-mongo');
 
 const buildMongoExecute = (query) => {
   return connectMongo().then(buildQueryMethods(query));
