+++ lib/drivers/mongo/crud.js
@@ -14,15 +14,34 @@ const MongoCrud = (engine, model) => {
     return _executeCount(query);
   };
 
+  const find = query => _find(query).then(_respondCamelized);
+
+  const findOne = query => _findOne(query).then(_respondCamelized);
+
+  const remove = query => _executeRemove(query).then(_respondCamelized);
+
+  const update = (query, data) => {
+    return _executeUpdate(snakeobj(query), snakeobj(data))
+       .then(_respondCamelized);
+  };
+
+  const insert = (data = {}) => {
+    return _executeInsert(snakeobj(data))
+      .then(_respondInsert)
+      .catch(_handleInsertError);
+  };
+
+  const setPrimaryKey = () => {};
+
+  const setSingleUnique = () => {};
+
+  const setCombinedUnique = () => {};
+
   const _executeCount = query => {
     const mongoQuery = transpiler.count(query);
     return buildMongoExecute(mongoQuery)
            .then(({count}) => count(collection));
   };
-
-  const find = query => _find(query).then(_respondFind);
-
-  const findOne = query => _findOne(query).then(_respondFindOne);
 
   const _find = query => {
     const mongoQuery = transpiler.select(query);
@@ -36,15 +55,7 @@ const MongoCrud = (engine, model) => {
           .then(({findOne}) => findOne(collection));
   };
 
-  const _respondFindOne = result => camelobj(result);
-
-  const _respondFind = data => camelobj(data);
-
-  const insert = (data = {}) => {
-    return _executeInsert(snakeobj(data))
-      .then(_respondSingle)
-      .catch(_handleInsertError);
-  };
+  const _respondCamelized = result => camelobj(result);
 
   const _handleInsertError = err => {
     throw CantInsertRecordError(err.message);
@@ -56,22 +67,11 @@ const MongoCrud = (engine, model) => {
        .then(({insert}) => insert(collection));
   };
 
-  const _respondRemove = data => data;
-
-  const remove = query => _executeRemove(query).then(_respondRemove);
-
   const _executeRemove = query => {
     const mongoQuery = transpiler.remove(query);
     return buildMongoExecute(mongoQuery)
            .then(({remove}) => remove(collection));
   };
-
-  const update = (query, data) => {
-    return _executeUpdate(snakeobj(query), snakeobj(data))
-       .then(_respondUpdate);
-  };
-
-  const _respondUpdate = data => camelobj(data);
 
   const _executeUpdate = (query, data) => {
     const mongoQuery = transpiler.update(query, data);
@@ -79,16 +79,10 @@ const MongoCrud = (engine, model) => {
        .then(({update}) => update(collection));
   };
 
-  const _respondSingle = rows => {
+  const _respondInsert = rows => {
     const result = rows.ops.pop();
     return result ? camelobj(toPlainObject(result)) : result;
   };
-
-  const setPrimaryKey = () => {};
-
-  const setSingleUnique = () => {};
-
-  const setCombinedUnique = () => {};
 
   return {
     findOne, find, count, insert, update, remove,
